Update to hybrid with small breaking changes
This is the initial implementation of the hybrid version of iron-lazy-pages.
Please test this pull request by installing `TimvdLippe/iron-lazy-pages#2.0-preview`.
It should work with Polymer 1.7 as well as Polymer 2.0-preview.

There are several breaking changes when upgrading from 1.0 to 2.0:

`iron-lazy-page` has been removed.`

There are two reasons for this. First of all type extension is not supported by all major browsers.
It is therefore not possible to rely on `<template is="...">` extensions.
Secondly, upgrading from `iron-pages` to `iron-lazy-pages` was quite intrusive
and the template syntax was confusing to users.

To upgrade, see the following example of before and after:

Before:
```html
<iron-lazy-pages attr-for-selected="data-route">
  <template is="iron-lazy-page" data-route="foo" path="foo.html">
    Foo page
  </template>
  <template is="iron-lazy-page" data-route="bar" path="bar.html">
    <bar-page></bar-page>
  </template>
</iron-lazy-pages>
```
After:
```html
<iron-lazy-pages attr-for-selected="data-route">
  <span data-route="foo" data-path="foo.html">Foo</span>
  <bar-page data-route="bar" data-path="bar.html"></bar-page>
</iron-lazy-pages>
```
As you can see, the templates are removed. Instead of `path` on `iron-lazy-pages`,
every page must now set `data-path` instead. This also removes compatibility for
plain text nodes (e.g. previously you could add plain text in a `iron-lazy-page`).
You must wrap the text in a `<span>` or `<div>` to once again display.

Web-animations is removed as dependency from `neon-lazy-pages`

If you are using `neon-lazy-pages` animations will not work out-of-the-box.
You must load `web-animations-js` yourself once in your application.
To do so add `"web-animations-js": "web-animations/web-animations-js#^2.2.0"` to your bower.json.
Then include `<script src="../../web-animations-js/web-animations-next-lite.min.js"></script>` in the head of your `index.html`.